requirements.txt
Django==3.1.2
django-crispy-forms==1.9.2

google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
